# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SmartHomeConfigResponse": {
        "properties": {
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Status"
          },
          "token_configured": {
            "default": true,
            "title": "Token Configured",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "base_url",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "title": "SmartHomeConfigResponse",
        "type": "object"
      },
      "SmartHomeConfigTestRequest": {
        "properties": {
          "base_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "token": {
            "anyOf": [
              {
                "maxLength": 4096,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          }
        },
        "title": "SmartHomeConfigTestRequest",
        "type": "object"
      },
      "SmartHomeConfigTestResponse": {
        "properties": {
          "home_assistant_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Assistant Version"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "message"
        ],
        "title": "SmartHomeConfigTestResponse",
        "type": "object"
      },
      "SmartHomeConfigUpdate": {
        "properties": {
          "base_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "token": {
            "anyOf": [
              {
                "maxLength": 4096,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          }
        },
        "required": [
          "base_url"
        ],
        "title": "SmartHomeConfigUpdate",
        "type": "object"
      },
...
      "SmartHomeServiceRequest": {
        "properties": {
          "entity_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Entity Id",
            "type": "string"
          },
          "service": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Service",
            "type": "string"
          },
          "service_data": {
            "additionalProperties": true,
            "title": "Service Data",
            "type": "object"
          }
        },
        "required": [
          "entity_id",
          "service"
        ],
        "title": "SmartHomeServiceRequest",
        "type": "object"
      },
      "SmartHomeServiceResponse": {
        "properties": {
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "service": {
            "title": "Service",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "entity_id",
          "service",
          "message"
        ],
        "title": "SmartHomeServiceResponse",
        "type": "object"
      },
...
      "UserConfigCreate": {
        "description": "Schema for creating user configuration.",
        "properties": {
          "data_retention_days": {
            "default": 365,
            "description": "Data retention period in days",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Data Retention Days",
            "type": "integer"
          },
          "feishu_webhook_url": {
            "default": "",
            "description": "Feishu webhook URL for notifications",
            "title": "Feishu Webhook Url",
            "type": "string"
          }
        },
        "title": "UserConfigCreate",
        "type": "object"
      },
      "UserConfigResponse": {
        "description": "Schema for user configuration response.",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "data_retention_days": {
            "default": 365,
            "title": "Data Retention Days",
            "type": "integer"
          },
          "feishu_webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Feishu Webhook Url"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ],
        "title": "UserConfigResponse",
        "type": "object"
      },
      "UserConfigUpdate": {
        "description": "Schema for updating user configuration.",
        "properties": {
          "data_retention_days": {
            "anyOf": [
              {
                "maximum": 3650.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Retention Days"
          },
          "feishu_webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Feishu webhook URL",
            "title": "Feishu Webhook Url"
          }
        },
        "title": "UserConfigUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Mavra",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/config": {
      "get": {
        "description": "Get current user configuration, or return defaults if not set.",
        "operationId": "config_get_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "config"
        ]
      },
      "patch": {
        "description": "Partial update user configuration (create if not exists).",
        "operationId": "config_update_config_partial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config Partial",
        "tags": [
          "config"
        ]
      },
      "post": {
        "description": "Create or update user configuration.",
        "operationId": "config_create_or_update_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Or Update Config",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/v1/smart-home/config": {
      "get": {
        "operationId": "smart_home_get_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartHomeConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "smart-home"
        ]
      },
      "put": {
        "operationId": "smart_home_update_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartHomeConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartHomeConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config",
        "tags": [
          "smart-home"
        ]
      }
    },
    "/api/v1/smart-home/config/test": {
      "post": {
        "operationId": "smart_home_test_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartHomeConfigTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartHomeConfigTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Config",
        "tags": [
          "smart-home"
        ]
      }
    },
...
    "/api/v1/smart-home/services/call": {
      "post": {
        "operationId": "smart_home_call_service",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartHomeServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartHomeServiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Call Service",
        "tags": [
          "smart-home"
        ]
      }
    },
...
  }
}